+++ UmfSparse.cc
@@ -6,7 +6,7 @@ UmfSparse::~UmfSparse() {
     if(Numeric) umfpack_di_free_numeric (&Numeric);
 }
 
-double& UmfSparse::operator()(int i, int j) {
+double& UmfSparse::operator()(unsigned int i, unsigned int j) {
     auto p = pair<int,int>(i,j);
     auto it = index.find(p);
     if(it != index.end()) return Ax[it->second];
@@ -20,7 +20,7 @@ double& UmfSparse::operator()(int i, int j) {
     return Ax.back();
 }
 
-double UmfSparse::operator()(int i, int j) const {
+double UmfSparse::operator()(unsigned int i, unsigned int j) const {
     auto it = index.find(pair<int,int>(i,j));
     if(it != index.end()) return Ax[it->second];
     return 0;
@@ -60,13 +60,13 @@ void UmfSparse::sort() {
     vector<int> newAi(Ai.size()), newAj(Aj.size());
     vector<double> newAx(Ax.size());
     
-    int i = 0;
-    int prevrow = 0;
+    unsigned int i = 0;
+    unsigned int prevrow = 0;
     for(auto it = index.begin(); it != index.end(); it++) {
         newAi[i] = Ai[it->second];
         newAj[i] = Aj[it->second];
         newAx[i] = Ax[it->second];
-        while(newAi[i] > prevrow) Ap[++prevrow] = i;
+        while(newAi[i] > (int)prevrow) Ap[++prevrow] = i;
         it->second = i++;
     }
     while(prevrow < mmax) Ap[++prevrow] = i;
